+++ tests/src/index.ts
@@ -13,6 +13,10 @@ import * as setFarmOwner from "./endpoints/set-farm-owner";
 import * as createFarmer from "./endpoints/create-farmer";
 import * as closeFarmer from "./endpoints/close-farmer";
 import * as startFarming from "./endpoints/start-farming";
+import * as whitelistFarmForCompounding from "./endpoints/whitelist-farm-for-compouding";
+import * as dewhitelistFarmForCompounding from "./endpoints/dewhitelist-farm-for-compounding";
+import * as compoundSameFarm from "./endpoints/compound-same-farm";
+import * as compoundAcrossFarms from "./endpoints/compound-across-farms";
 import * as stopFarming from "./endpoints/stop-farming";
 import * as updateEligibleHarvest from "./endpoints/update-eligible-harvest";
 import * as claimEligibleHarvest from "./endpoints/claim-eligible-harvest";
@@ -29,10 +33,14 @@ describe("farming", () => {
   setTokensPerSlot.test();
   createFarmer.test();
   startFarming.test();
+  closeFarmer.test();
+  whitelistFarmForCompounding.test();
+  dewhitelistFarmForCompounding.test();
+  compoundSameFarm.test();
+  compoundAcrossFarms.test();
   stopFarming.test();
   updateEligibleHarvest.test();
   claimEligibleHarvest.test();
-  closeFarmer.test();
 
   before("airdrop SOL to provider wallet", async () => {
     await airdrop(provider.wallet.publicKey);